feat(i18n): add `replyLocalized`
Also fixed some of the TSDoc and moved the interfaces to the `types` file.
